[LibOS] Rework signal handling and syscall emulation
Change log (most important only):
- unify cpu context structures - now we have only one version -
  `PAL_CONTEXT` - which is shared between LibOS and Pals and it should
  depend only on the host architecture (not OS),
- syscalls emulation changed:
  - dedicated LibOS stack is now used for syscalls emulation,
  - removed one indirection level in syscalls table - now it stores
    `shim_do_*` function directly,
- signal handling - completely rewritten:
  - all signal queues use proper locking schemes now,
  - signals are handled *only* when returning to the user app from LibOS
    or Pal,
  - nested signals are now possible,
  - the app is allowed to jump out of signal handler with the same
    sematics as on normal Linux,
  - signal altstack is now fully supported,
  - syscall restarting is now supported.

+++ LibOS/shim/src/sys/shim_alarm.c
@@ -27,7 +27,7 @@ static void signal_alarm(IDTYPE caller, void* arg) {
     }
 }
 
-int shim_do_alarm(unsigned int seconds) {
+long shim_do_alarm(unsigned int seconds) {
     uint64_t usecs = 1000000ULL * seconds;
 
     int64_t ret = install_async_event(NULL, usecs, &signal_alarm, NULL);
@@ -66,8 +66,8 @@ static void signal_itimer(IDTYPE target, void* arg) {
 #define ITIMER_REAL 0
 #endif
 
-int shim_do_setitimer(int which, struct __kernel_itimerval* value,
-                      struct __kernel_itimerval* ovalue) {
+long shim_do_setitimer(int which, struct __kernel_itimerval* value,
+                       struct __kernel_itimerval* ovalue) {
     if (which != ITIMER_REAL)
         return -ENOSYS;
 
@@ -114,7 +114,7 @@ int shim_do_setitimer(int which, struct __kernel_itimerval* value,
     return 0;
 }
 
-int shim_do_getitimer(int which, struct __kernel_itimerval* value) {
+long shim_do_getitimer(int which, struct __kernel_itimerval* value) {
     if (which != ITIMER_REAL)
         return -ENOSYS;
