+++ LibOS/shim/src/sys/shim_eventfd.c
@@ -54,7 +54,7 @@ static int create_eventfd(PAL_HANDLE* efd, unsigned count, int flags) {
     return 0;
 }
 
-int shim_do_eventfd2(unsigned int count, int flags) {
+long shim_do_eventfd2(unsigned int count, int flags) {
     int ret = 0;
     struct shim_handle* hdl = get_new_handle();
 
@@ -86,6 +86,6 @@ out:
     return ret;
 }
 
-int shim_do_eventfd(unsigned int count) {
+long shim_do_eventfd(unsigned int count) {
     return shim_do_eventfd2(count, 0);
 }
